# Appointments.Api.Client/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Appointments.Api.Client",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
...
    "/api/tenant/tenants/{id}": {
      "get": {
        "tags": [
          "Tenants"
        ],
        "operationId": "GetTenant",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantProfileResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Tenants"
        ],
        "operationId": "UpdateTenantProfile",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTenantProfileRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTenantProfileRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTenantProfileRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/tenant/tenants/{id}/branch-offices": {
      "post": {
        "tags": [
          "Tenants"
        ],
        "operationId": "CreateBranchOffice",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBranchOfficeRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBranchOfficeRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBranchOfficeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BranchOfficeCreatedResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Tenants"
        ],
        "operationId": "FindBranchOffices",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "pageIndex",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          },
          {
            "name": "sort",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filter",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BranchOfficeListResponsePagedResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/tenant/tenants/{id}/branch-offices/{branchOfficeId}": {
      "get": {
        "tags": [
          "Tenants"
        ],
        "operationId": "GetBranchOffice",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "branchOfficeId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BranchOfficeProfileResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Tenants"
        ],
        "operationId": "UpdateBranchOffice",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "branchOfficeId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBranchOfficeRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBranchOfficeRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBranchOfficeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Tenants"
        ],
        "operationId": "DeleteBranchOffice",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "branchOfficeId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AddressModel": {
        "type": "object",
        "properties": {
          "coordinates": {
            "$ref": "#/components/schemas/CoordinatesModel"
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "BranchOfficeCreatedResponse": {
        "type": "object",
        "properties": {
          "branchOfficeId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
      "BranchOfficeListResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdBy": {
            "type": "string",
            "nullable": true
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "updatedBy": {
            "type": "string",
            "nullable": true
          },
          "tenantId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "address": {
            "$ref": "#/components/schemas/AddressModel"
          }
        },
        "additionalProperties": false
      },
      "BranchOfficeListResponsePagedResult": {
        "type": "object",
        "properties": {
          "pageIndex": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "totalCount": {
            "type": "integer",
            "format": "int64"
          },
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BranchOfficeListResponse"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "BranchOfficeProfileResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdBy": {
            "type": "string",
            "nullable": true
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "updatedBy": {
            "type": "string",
            "nullable": true
          },
          "tenantId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "address": {
            "$ref": "#/components/schemas/AddressModel"
          },
          "contacts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SocialMediaContactModel"
            },
            "nullable": true
          },
          "schedule": {
            "$ref": "#/components/schemas/WeeklyScheduleModel"
          }
        },
        "additionalProperties": false
      },
      "CoordinatesModel": {
        "type": "object",
        "properties": {
          "longitude": {
            "type": "integer",
            "format": "int32"
          },
          "latitude": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "CreateBranchOfficeRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "address": {
            "$ref": "#/components/schemas/AddressModel"
          },
          "contacts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SocialMediaContactModel"
            },
            "nullable": true
          },
          "schedule": {
            "$ref": "#/components/schemas/WeeklyScheduleModel"
          }
        },
        "additionalProperties": false
      },
      "DailySchedule": {
        "type": "object",
        "properties": {
          "hours": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DateRange"
            },
            "nullable": true
          },
          "disabled": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "DailyScheduleModel": {
        "type": "object",
        "properties": {
          "hours": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DateRangeModel"
            },
            "nullable": true
          },
          "disabled": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "DateRange": {
        "type": "object",
        "properties": {
          "startDate": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          },
          "endDate": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          },
          "disabled": {
            "type": "boolean",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "DateRangeModel": {
        "type": "object",
        "properties": {
          "startDate": {
            "type": "string",
            "format": "date-time"
          },
          "endDate": {
            "type": "string",
            "format": "date-time"
          },
          "disabled": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "SocialMediaContact": {
        "type": "object",
        "properties": {
          "type": {
            "$ref": "#/components/schemas/SocialMediaType"
          },
          "otherType": {
            "type": "string",
            "nullable": true
          },
          "contact": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "SocialMediaContactModel": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "otherType": {
            "type": "string",
            "nullable": true
          },
          "contact": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "SocialMediaType": {
        "enum": [
          0,
          1,
          2,
          3,
          4
        ],
        "type": "integer",
        "format": "int32"
      },
      "TenantProfileResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdBy": {
            "type": "string",
            "nullable": true
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "updatedBy": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "slogan": {
            "type": "string",
            "nullable": true
          },
          "urlId": {
            "type": "string",
            "nullable": true
          },
          "logo": {
            "type": "string",
            "nullable": true
          },
          "contacts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SocialMediaContact"
            },
            "nullable": true
          },
          "schedule": {
            "$ref": "#/components/schemas/WeeklySchedule"
          }
        },
        "additionalProperties": false
      },
      "TenantUrlId": {
        "type": "object",
        "properties": {
          "value": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "UpdateBranchOfficeRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "address": {
            "$ref": "#/components/schemas/AddressModel"
          },
          "contacts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SocialMediaContactModel"
            },
            "nullable": true
          },
          "schedule": {
            "$ref": "#/components/schemas/WeeklyScheduleModel"
          }
        },
        "additionalProperties": false
      },
      "UpdateTenantProfileRequest": {
        "type": "object",
        "properties": {
          "updatedBy": {
            "type": "string",
            "nullable": true
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "slogan": {
            "type": "string",
            "nullable": true
          },
          "urlId": {
            "$ref": "#/components/schemas/TenantUrlId"
          },
          "contacts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SocialMediaContact"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "WeeklySchedule": {
        "type": "object",
        "properties": {
          "monday": {
            "$ref": "#/components/schemas/DailySchedule"
          },
          "tuesday": {
            "$ref": "#/components/schemas/DailySchedule"
          },
          "wednesday": {
            "$ref": "#/components/schemas/DailySchedule"
          },
          "thursday": {
            "$ref": "#/components/schemas/DailySchedule"
          },
          "friday": {
            "$ref": "#/components/schemas/DailySchedule"
          },
          "saturday": {
            "$ref": "#/components/schemas/DailySchedule"
          },
          "sunday": {
            "$ref": "#/components/schemas/DailySchedule"
          }
        },
        "additionalProperties": false
      },
      "WeeklyScheduleModel": {
        "type": "object",
        "properties": {
          "monday": {
            "$ref": "#/components/schemas/DailyScheduleModel"
          },
          "tuesday": {
            "$ref": "#/components/schemas/DailyScheduleModel"
          },
          "wednesday": {
            "$ref": "#/components/schemas/DailyScheduleModel"
          },
          "thursday": {
            "$ref": "#/components/schemas/DailyScheduleModel"
          },
          "friday": {
            "$ref": "#/components/schemas/DailyScheduleModel"
          },
          "saturday": {
            "$ref": "#/components/schemas/DailyScheduleModel"
          },
          "sunday": {
            "$ref": "#/components/schemas/DailyScheduleModel"
          }
        },
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}